Added useId; reconstructed utils/ #804
