+++ src/Theming/index.js
@@ -12,4 +12,4 @@ import ThemeContext from './ThemeContext';
 const { Consumer : ThemeConsumer, Provider : ThemeProvider } = ThemeContext;
 export { ThemeConsumer, ThemeProvider };
 
-export useTheme from '../utils/customHooks';
+export useTheme from '../utils';
